+++ src/main/java/com/example/serviceImple/UserCreateServiceImple.java
@@ -33,7 +33,7 @@ public class UserCreateServiceImple
 
         Role role =
                 roleRepository.findById(
-                        (int) user.getRole().getRoleId())
+                        user.getRole().getRoleId())
                 .orElse(null);
 
         List<Address> addresses =
